* Feature - Add a "Delivered" booking status, in order to flag when the customer has attended the booking * Feature - Add {price} notification tag to display the WC order item price bound to the booking * Tweak - Events dropped on the month view now start at the time set in the "Day begin" option * Tweak - Display shortcuts in calendar editor * Tweak - Change the "Refunded" label to "Refunded with coupon ..." if it was refunded with a coupon * Tweak - Add an option to delete all Booking Activities data on uninstall, in order to avoid accidental data deletion * Tweak - Set default snap frequency to 5 minutes instead of 30

+++ booking-activities.php
@@ -331,20 +331,23 @@ register_deactivation_hook( __FILE__, 'bookacti_deactivate' );
 
 /**
  * Uninstall Booking Activities
+ * @version 1.6.0
  */
 function bookacti_uninstall() {
-	//Deregister the hourly reccuring event
+	// Deregister the hourly reccuring event
 	wp_clear_scheduled_hook( 'bookacti_hourly_event' );
-
-	// Delete plugin settings
-	bookacti_delete_settings();
 	
 	// Delete notices acknowledgement
 	bookacti_reset_notices();
 	
-	// Drop tables and every Booking Activities Data
-	delete_option( 'bookacti_version' );
-	bookacti_drop_tables();
+	if( bookacti_get_setting_value( 'bookacti_general_settings', 'delete_data_on_uninstall' ) ) {
+		// Delete plugin settings
+		bookacti_delete_settings();
+
+		// Drop tables and every Booking Activities Data
+		delete_option( 'bookacti_version' );
+		bookacti_drop_tables();
+	}
 	
 	// Unset roles and capabilities
 	bookacti_unset_role_and_cap();
